molookup-related init operations have been modified, such that the modes of operation are ordered to have increasing frequencies.
A number of loose global varriables, including LAMbS_mo_count, have been concolidated into a single global structure called LAMbS_mo_struct. The deffinition of the structure and the extern deffinition of the structure varriable are in LAMbS_mo_lookup.h.

+++ pbs_module/LAMbS_mo.h
@@ -8,8 +8,7 @@ int     LAMbS_mo_init(int verbose);
 void    LAMbS_mo_uninit(void);
 void    LAMbS_mo_modelupdate_notify(void);
 
-extern int  LAMbS_current_moi;
+extern u64  LAMbS_current_moi;
 extern u64  LAMbS_current_instretirementrate;
 extern u64  LAMbS_current_instretirementrate_inv;
-
 #endif
